requirements.txt
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
numpy==1.24.2
opencv-python-headless==4.7.0.72
Pillow==9.4.0
redis==4.5.1
requests==2.28.2
six==1.16.0
spotipy==2.22.1
urllib3==1.26.14
python-dotenv==1.0.0